LangChain-Live-Session/pyproject.toml
[project]
name = "aillm3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain==1.2.0",
    "langchain-google-genai==4.1.2",
    "langgraph-checkpoint-postgres>=3.0.2",
    "pandas",
    "psycopg2-binary>=2.9.11",
    "python-dotenv==1.2.1",
    "sqlmodel>=0.0.14",
]
